+++ apps/core/feeds/urls.py
@@ -19,8 +19,8 @@ urlpatterns = [
         name="feed_sources",
     ),
     api_path(
-        "<str:feed_id>/sources/<str:source_id>",
-        views.FeedSourceDetailView.as_view(),
-        name="feed_source_detail",
+        "<str:feed_id>/items",
+        views.FeedItemsView.as_view(),
+        name="feed_items",
     ),
 ]